pyproject.toml
[project]
name = "langchain-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "arxiv>=2.4.1",
    "bs4>=0.0.2",
    "chromadb>=1.5.8",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "grandalf>=0.8",
    "huggingface>=0.0.1",
    "huggingface-hub>=1.7.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-google-vertexai>=3.2.2",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "langchainhub>=0.1.21",
    "langgraph>=1.1.3",
    "langserve>=0.3.3",
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "schema>=0.7.8",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "sse-starlette>=3.3.3",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
    "transformers>=5.3.0",
    "uvicorn>=0.42.0",
    "wikipedia>=1.4.0",
]
